requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-crispy-forms==1.14.0
django-phonenumber-field==6.4.0
greenlet==1.1.3
idna==3.3
msgpack==1.0.4
phonenumbers==8.12.54
pynvim==0.4.3
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
gunicorn==20.0.4
psycopg2-binary==2.8.6
